css fixes for CaptionedImage to be used throughout all banners, moved ContribuIcons into components, cleanup of duplicate copy images in res folder
