+++ src/Components/ContribIcons/res/ideaIcon.png
