+++ src/Pages/AboutProject/About.css
@@ -5,67 +5,28 @@
 /*
 BANNER
 */
-
 .aboutBanner {
-  width: 100vw;
-  background: url("../../res/mapbg.png");
-  padding-top: 151rem;
-  display: flex;
-}
-
-.aboutBanner_button {
-  width: 500rem;
-  height: 500rem;
-}
-
-.aboutBanner_card {
-  width: 834rem;
-  height: 358rem;
-  min-height: 358rem;
-  background-color: white;
-  border-left: 12rem solid var(--colorDeepRed);
-  padding: 61rem;
-  padding-top: 58rem;
-  margin-top: 52rem;
-}
-
-.aboutBanner_card p {
-  font-family: var(--fontNeoBulletin);
-  font-size: 30rem;
-  font-weight: bold;
-  width: 760rem;
-}
-
-.aboutBanner_credit {
-  min-width: 323rem;
-  height: 84rem;
-  background-color: var(--colorDarkBlue);
-  color: var(--colorBeige);
-  font-size: 36rem;
-  font-family: var(--fontBabasNeue);
   display: flex;
   align-items: center;
-  justify-content: center;
-  right: 0;
-  margin-left: -200rem;;
-  margin-top: 577rem;
-  margin-right: 400rem;
-  width: fit-content;
-  text-align: right;
-  padding: 12rem;
+  justify-content: space-around;
+  background: url("../../res/mapbg.png");
 }
 
-.aboutBanner_credit p {
-  width: 323rem;
-  text-align: center;
+/*
+Control the individal space for each item in the banner
+*/
+.aboutBanner > img {
+  width: 25%;
 }
 
-.aboutBanner_chick {
-  width: 537rem;
-  height: 764rem;
-  margin-left: -400rem;
-  margin-bottom: -151rem;
-  margin-top: -100rem;
+.aboutBanner > .lcard {
+  width: 50%;
+  margin: 2em 0;
+}
+
+.aboutBanner > .captionedImg {
+  width: 25%;
+  margin-left: 4em;
 }
 
 .aboutHead > h1 {
