+++ src/Pages/HowToContribute/HowToContribute.css
@@ -14,42 +14,27 @@ html {
 
 .contributeBanner {
   display: flex;
-  width: 100%;
   align-items: center;
   justify-content: space-around;
   background: url("../../res/mapbg.png");
 }
 
+/*
+Control the individal space for each item in the banner
+*/
 .contributeBanner > img {
   width: 25%;
-  
 }
 
 .contributeBanner > .lcard {
-  width: 35%;
+  width: 50%;
   margin: 2em 0;
 }
 
 .contributeBanner > .captionedImg {
-  width: 30%;
-  align-self: flex-end;
-  /* margin-bottom: -3rem; */
-  margin-right: -40rem;
+  width: 25%;
+  margin-left: 4em;
 }
-
-.captionedImg{
-  position: relative;
-  width: 100%;
-  /* height: 770rem; */
-}
-
-.captionedImg p{
-  position: absolute;
-  top: 785rem;
-  left: 0;
-}
-
-
 
 /* 
     HOW TO GET INVOLVED BANNER TEXT SOCIAL MEDIA
